pyproject.toml
[project]
name = "aerohawkspython"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikita Nasiedkin",email = "nkitanas@proton.me"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gpiod (>=2.4.0,<3.0.0)",
    "mpu6050-raspberrypi (>=1.2,<2.0)",
    "lorarf (>=1.4.0,<2.0.0)",
    "smbus (>=1.1.post2,<2.0)",
    "spidev (>=3.8,<4.0)"
]

[tool.poetry]
packages = [{include = "aerohawkspython", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
